requirements.txt
anyio==4.3.0
astroid==3.3.4
asyncio==3.4.3
black==24.8.0
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
construct==2.10.68
construct-typing==0.5.6
dill==0.3.9
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
isort==5.13.2
jsonalias==0.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pyaes==1.6.1
pylint==3.3.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.1
solana==0.32.0
solders==0.20.0
termcolor==2.4.0
TgCrypto==1.2.5
tomlkit==0.13.2
types-cachetools==4.2.10
typing_extensions==4.10.0
urllib3==2.2.1
uvloop==0.19.0
websockets==11.0.3
